Add a standalone example program
This is a helper library so it can't do much by itself. But, since it
was requested a dummy standalone program was added. The purpose of the
program is just to illustrate how the library can be used in a form that
is more convenient to hack on (as opposed to the tests). The program
simply prints a grid for every byte in the word in the uncompressed
image.